+++ src/home/PurchaseCard.js
@@ -1,22 +1,12 @@
 import React, { useState } from 'react';
 
-import { Card, CardContent, CardActions, Button, CardMedia, TextField } from '@material-ui/core';
-import { Dialog, DialogActions, DialogContent, DialogContentText, DialogTitle } from '@material-ui/core'
+import { Card, CardContent, CardActions, Button, CardMedia} from '@material-ui/core';
 
-const HomeDialogField = ({id, label, type, value, onChange, autoFocus, rows, multiline}) => <TextField className='home-dialog-form-field' {...{id, label, type, onChange, value, autoFocus, rows, multiline}}  />
-const NameField = ({value, onChange}) => <HomeDialogField id='purchase-name' label='Name' type='text' {...{onChange, value}} autoFocus />
-const EmailField = ({value, onChange}) => <HomeDialogField id='purchase-email' label='Email' type='email' {...{onChange, value}} />
-const PhoneField = ({value, onChange}) => <HomeDialogField id='purchase-phone' label='Phone Number' type='number' {...{onChange, value}} />
-const CommentField = ({value, onChange}) => <HomeDialogField id='purchase-comments' label='Comments' {...{onChange, value}} multiline rows={5} />
-
+import PurchaseForm from './PurchaseForm'
 
 
 export const PurchaseCard = ({header,price,description,img}) => {
     const [dialogOpen, setDialog] = useState(false)
-    const [name, setName] = useState('')
-    const [email, setEmail] = useState('')
-    const [phone, setPhone] = useState('')
-    const [comment, setComment] = useState('')
     return (
         <Card className='purchase-card home-display-card'>
             <div className='purchase-card-header home-container-header'> {header} </div>
@@ -34,57 +24,11 @@ export const PurchaseCard = ({header,price,description,img}) => {
                     Buy
                 </Button>
             </CardActions>
-            <Dialog
-                open={dialogOpen}
-                onClose={()=> setDialog(false)}
-            >
-                <DialogTitle>Purchase Inquiry</DialogTitle>
-                <DialogContent>
-                    <DialogContentText>
-                        Please fill in your information below and we will get back to you as soon as we can.
-                    </DialogContentText>
-                    <NameField onChange={e => setName(e.target.value)} value={name} />
-                    <br/>
-                    <EmailField onChange={e => setEmail(e.target.value)} value={email}/>
-                    <br/>
-                    <PhoneField onChange={e => setPhone(e.target.value)} value={phone} />
-                    <br/>
-                    <CommentField onChange={e => setComment(e.target.value)} value={comment}/>
-                </DialogContent>
-                <DialogActions>
-                    <Button onClick={() => setDialog(false)} color='secondary'> Cancel </Button>
-                    <Button onClick={() => {
-                      handleOk(name, email, phone, comment, description)
-                      setDialog(false)
-                      }} color='primary'> Submit </Button>
-                </DialogActions>
-            </Dialog>
+            <PurchaseForm  description={description} dialogOpen={dialogOpen} />
         </Card>
     )
 }
 
-const handleOk = async function (name, email, phone, comment, description, setDialog) {
-  comment = `${comment}
-  DESCRIPTION: ${description} ` // will add support  for explicit part number field next
-  const fieldValues = {
-    name, email, phone, comment
-  }
-  if(!email && !phone) {
-    alert('email or phone is required')
-  }
-  const response = await fetch('https://yegwheel2019as.azurewebsites.net/api/inquiry',{
-    method:'POST',
-    cache: 'no-cache',
-    headers: {
-      'Content-Type':'application/json'
-    },
-    body: JSON.stringify(fieldValues)
-  })
-  const json = await response.json()
-  const {id} = json
-  if(id) alert(`thank you for your request, the id of the request is ${id}; we will be in touch.`)
-  else alert(`Our apologinies, we couldn't process your request, please try again later.`)
-}
 
 
 export default PurchaseCard;
